requirements.txt
astroid==2.2.5
backcall==0.1.0
decorator==4.3.2
gunicorn==19.9.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.15
jedi==0.13.3
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.3.1
six==1.12.0
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
WebOb==1.8.5
wrapt==1.11.1
